currency update fixed
