+++ src/app.ts
@@ -5,6 +5,8 @@ import apiErrorHandler from "./middleware/apiErrorHandler";
 import exchangeRateRoutes from "./routes/exchangeRate";
 import userRoutes from "./routes/user";
 import chatGptRoutes from "./routes/chatGptRoutes";
+import { writeFiles } from "./services/exchangeRate";
+import schedule from "node-schedule";
 
 const app = express();
 
@@ -16,5 +18,17 @@ app.use("/users", userRoutes);
 app.use("/chatgpt", chatGptRoutes);
 
 app.use(apiErrorHandler);
+
+writeFiles();
+
+const rule = new schedule.RecurrenceRule();
+rule.dayOfWeek = [new schedule.Range(1, 5)];
+rule.hour = 17;
+rule.minute = 0;
+
+const dailyUpdate = schedule.scheduleJob(rule, function () {
+  console.log("db update run");
+  writeFiles();
+});
 
 export default app;
